introduced _scrollable to avoid re-initializing
With changes introduced in Ember 1.13.0, route refreshing doesn't call didInsertElement anymore,
but is still re-initializing components. Since 'scrollable' was both a parameter of the
ember-infinity component and the property storing the jQuery wrapper initialized in didInsertElement,
it was reset to the initial parameter value.

By storing the jQuery wrapped scrollable DOM element in _scrollable, we avoid it getting re-initialized.

With this change we can also drop the hacky fix introduced by myself earlier, which involved
doing all the setup work, normally done in didInsertElement, in didRender as well. This introduced
a bunch of other issues which are gone as of this commit as well.